File messageCodes is moved into Utils folder from config Folder
